requirements.txt
pyyaml~=6.0.1
opencv-python~=4.8.1.78
numpy~=1.23.0
torch~=1.12.1+cu102
torchvision~=0.13.1+cu102
pillow~=10.1.0
pyiqa~=0.1.7
timm~=0.9.2
wandb~=0.16.0
termcolor~=2.1.0
lightning~=2.0.4
thop~=0.1.1-2209072238
einops~=0.7.0